+++ cScripts/CavFnc/functions/players/fn_getRadioChannel.sqf
@@ -38,6 +38,13 @@ if (_radio == "ACRE_PRC343") then {
         case "APOLLO-5": {_channel = 1};
         case "APOLLO-6": {_channel = 1};
         
+        case "ATLAS-1": {_channel = 2};
+        case "ATLAS-2": {_channel = 3};
+        case "ATLAS-3": {_channel = 1};
+        case "ATLAS-4": {_channel = 1};
+        case "ATLAS-5": {_channel = 1};
+        case "ATLAS-6": {_channel = 1};
+        
         //Air
         case "PEGASUS-2": {_channel = 1};
 
@@ -122,6 +129,13 @@ if (_radio == "ACRE_PRC343") then {
         case "APOLLO ONE": {_channel = 1};
         case "APOLLO FIVE": {_channel = 1};
         case "APOLLO SIX": {_channel = 1};
+        
+        case "ATLAS ONE": {_channel = 2};
+        case "ATLAS TWO": {_channel = 3};
+        case "ATLAS THREE": {_channel = 1};
+        case "ATLAS FOUR": {_channel = 1};
+        case "ATLAS FIVE": {_channel = 1};
+        case "ATLAS SIX": {_channel = 1};
         
         //Air
         case "PEGASUS TWO": {_channel = 1};
@@ -214,6 +228,13 @@ if (_radio == "ACRE_PRC343") then {
         case "APOLLO-5": {_channel = 15};
         case "APOLLO-6": {_channel = 15};
         
+        case "ATLAS-1": {_channel = 15};
+        case "ATLAS-2": {_channel = 15};
+        case "ATLAS-3": {_channel = 15};
+        case "ATLAS-4": {_channel = 15};
+        case "ATLAS-5": {_channel = 15};
+        case "ATLAS-6": {_channel = 15};
+        
         //Air
         case "PEGASUS-2": {_channel = 2};
 
@@ -299,6 +320,13 @@ if (_radio == "ACRE_PRC343") then {
         case "APOLLO ONE": {_channel = 15};
         case "APOLLO FIVE": {_channel = 15};
         case "APOLLO SIX": {_channel = 15};
+        
+        case "ATLAS ONE": {_channel = 15};
+        case "ATLAS TWO": {_channel = 15};
+        case "ATLAS THREE": {_channel = 15};
+        case "ATLAS FOUR": {_channel = 15};
+        case "ATLAS FIVE": {_channel = 15};
+        case "ATLAS SIX": {_channel = 15};
         
         //Air
         case "PEGASUS TWO": {_channel = 2};
